fix: table users cloums id data type is bigint

+++ src/main/resources/db/migration/V1__Base_version.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users(
-    id int primary key not null,
+    id bigint primary key not null,
     username varchar(100) not null,
     password varchar(100) not null
 )
